+++ lib/models/money_objects/money_object.dart
@@ -11,6 +11,11 @@ class MoneyObject<C> {
   /// All object must have a unique identified
   int get uniqueId => -1;
 
+  /// State of any and all object instances
+  /// to indicated any alteration to the data set of the users
+  /// to reflect on the customer CRUD actions [Create|Rename|Update|Delete]
+  ChangeType change = ChangeType.none;
+
   MoneyObject<C> fromJson(final MyJson row) {
     return MoneyObject<C>();
   }
@@ -28,4 +33,15 @@ class MoneyObject<C> {
   /// SubTitle    |
   ///
   Widget Function()? buildListWidgetForSmallScreen = () => const Text('Small screen content goes here');
+}
+
+enum ChangeType {
+  none,
+  changed,
+  inserted,
+  deleted,
+  reloaded,
+  rebalanced,
+  childChanged,
+  transientChanged,
 }